backend/requirements.txt
# Web Push Notifications
pywebpush==2.0.0
py-vapid==1.8.2
apscheduler==3.10.4
firebase-admin==6.2.0

# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.1

# Database (for future migration from JSON)
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL driver
# pymysql==1.1.0  # MySQL driver (optional)

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality & Linting
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1
pylint==3.0.3

# Logging & Monitoring
python-json-logger==2.0.7
# sentry-sdk==1.38.0  # Error tracking (optional)

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2

# Development
ipython==8.18.1
ipdb==0.13.13
